docs: update htmx overview and links
Added a detailed description of HTMX in the overview section along with an image and a YouTube video. Updated the "Getting Started" button link on the homepage, as well as the "Read the docs" link. Also added new HTML tags and classes to hugo_stats.json while removing unused ones. Lastly, included a new image file for HTMX in static/images directory.