docs: Update documentation for RzPage and RzPartial usage
The documentation for `RzPage` and `RzPartial` has been updated. The reference to the use of cascading values to pass `RzViewContext` for accessing app-wide context information has been removed. Also, the requirement of providing a `ViewContext` when rendering `RzPartial` through services or controllers is no longer necessary. In addition, the 'ViewContext' property in Minimal APIs overview is now marked as obsolete.