docs: update method names in htmx configuration
The documentation has been updated to reflect changes in the method names used for adding custom HTMX configurations. The `AddRizzy()` and `WithHtmxConfiguration()` methods have been replaced with `AddHtmx()` and `Configure<HtmxConfig>()` respectively. This change simplifies the process of adding default and named configurations, as well as configuring nonce values in HTMX.